 requirements.txt
python-telegram-bot==20.3
playwright==1.39.0
beautifulsoup4==4.12.2
jdatetime==4.1.0
python-dotenv==1.0.0
pytest==7.4.0